Images handling implemented in all layers

+++ Backend/src/main/java/pw/react/backend/service/FlatImageService.java
@@ -1,5 +1,7 @@
 package pw.react.backend.service;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
@@ -7,17 +9,20 @@ import org.springframework.web.multipart.MultipartFile;
 import pw.react.backend.appException.InvalidFileException;
 import pw.react.backend.appException.ResourceNotFoundException;
 import pw.react.backend.dao.FlatImageRepository;
-import pw.react.backend.model.CompanyLogo;
 import pw.react.backend.model.FlatImage;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
+import static java.util.stream.Collectors.joining;
 
 @Service
 public class FlatImageService implements ImageService
 {
-
+    private final Logger logger = LoggerFactory.getLogger(ImageService.class);
     FlatImageRepository repository;
 
     @Autowired
@@ -40,9 +45,7 @@ public class FlatImageService implements ImageService
             {
                 if (fileName.contains(".."))
                     throw new InvalidFileException("Sorry! Filename contains invalid path sequence " + fileName);
-                FlatImage newImage = new FlatImage(fileName, file.getContentType(), flatId, false, file.getBytes());
-//                repository.findByFlatId(flatId).ifPresent(flatImage -> newImage.setId(flatImage.getId()));
-//                repository.findByCompanyId(companyId).ifPresent(companyLogo -> newCompanyLogo.setId(companyLogo.getId()));
+                FlatImage newImage = new FlatImage(fileName, file.getContentType(), flatId, file.getBytes());
                 saved.add(repository.save(newImage));
             } catch (IOException ex) {
                 throw new InvalidFileException("Could not store file " + fileName + ". Please try again!", ex);
@@ -54,16 +57,37 @@ public class FlatImageService implements ImageService
     @Override
     public List<FlatImage> getFlatImages(long flatId)
     {
-        List<FlatImage> list = repository.getAllByFlatId(flatId);
-        if (list.isEmpty())
-            throw new ResourceNotFoundException("No file found for flat with id " + flatId);
-        return list;
+        return repository.getAllByFlatId(flatId);
     }
 
     @Override
-    public boolean deleteFlatImage(long flatId, String imageId)
+    public Optional<FlatImage> getFirstImage(long flatId)
     {
-        repository.deleteById(imageId);
-        logger.info("Logo for the company with id {} deleted.", companyId);
+        return repository.getFirstByFlatId(flatId);
     }
+
+    @Override
+    public void deleteFlatImage(long flatId, String imageId)
+    {
+        if (repository.findById(imageId).isPresent()) {
+            repository.deleteById(imageId);
+            logger.info("Logo for the company with id {} deleted.", flatId);
+        }
+        throw new ResourceNotFoundException("No file found for flat with id " + flatId);
+    }
+
+    @Override
+    public void deleteImagesForFlat(long flatId)
+    {
+        repository.deleteByFlatId(flatId);
+    }
+
+    @Override
+    public void deleteRemovedImages(List<FlatImage> updatedImages)
+    {
+        var imageIds = updatedImages.stream().map(FlatImage::getId).collect(Collectors.toList());
+        repository.deleteByIdNotIn(imageIds);
+        logger.info(String.format("Removing ids no in %s", updatedImages.stream().map(FlatImage::getId).collect(joining(","))));
+    }
+
 }
